Initial commit! 🎉
This commits an initial working [Cypress based](https://www.cypress.io/) acceptance test framework.

It is very much based on our [SROC acceptance test project](https://github.com/DEFRA/sroc-acceptance-tests) which is used to test the [SROC Tactical Charging Module](https://github.com/DEFRA/sroc-tcm-admin). Just like it, this project uses [Cucumber BDD](https://cucumber.io/) style tests as opposed to vanila action/assert (bear this in mind when looking for examples on the web!)

Differences are configuration comes from the `.env` files in `environments/`, rather than `.json` files. This is because the intent is `.env` will store credentials which **must** not be committed. It incorporates [cy-api](https://github.com/bahmutov/cy-api) which allows us to see the requests and responses in Cypress itself (normally you'd be watching the website under test). It also has the ability to generate a HTML report. It still uses the idea of [page objects](https://martinfowler.com/bliki/PageObject.html). However, as the API deals with 'endpoints' that's how we refer to them in the project.

The intent is to replace [sroc-cha-acceptance-tests](https://github.com/DEFRA/sroc-cha-acceptance-tests) which is based on [Postman](https://www.postman.com/). It also has the ability to generate a HTML report of the test results and we didn't want to lose this feature in the switch.

Finally, we have included some initial tests as a base to work from. They demonstrate `POST`, `PATCH` and `GET` requests. They also include polling an endpoint till an expected result is returned.