Add Dependabot and a GitHub CI workflow
We like to keep our projects up to date to ensure we don't fall to far behind. If we did there are features we might miss out on, fixes we're not aware of, and security patches we should apply (though a test suite we do run them in our environments).

We also like to know that we haven't broken anything when we make a change. So, as a standard we have a [GitHub Workflow](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions) that runs just to confirm the project still builds and can run a specific CI test.