Changes to allow the feature to be run in CI
We want to be able to run our CI feature in GitHub actions and only the CI feature. In doing this we've learnt that Cypress does an automatic check on the web service under test. If it cannot connect then it won't bother running the tests. This is a bit of a shame as our CI feature is intended to have no dependency on anything external but it's something we cannot work around.

Currently, the project is configured to default to selecting the 'local' environment if none is provided and this means it tries to connect to http://localhost:3000 in its startup check. So, rather than re-writing everything, the simplest solution is a CI .env file that uses a baseURL we know is simple and reliable (well done example.com!)

Finally, we add an npm script to `package.json` for running just the CI feature.