Add .nvmrc file
As a standard in our projects we add an [.nvmrc](https://github.com/nvm-sh/nvm#nvmrc) to specify the version of Node being used. This helps ensure consistency across the tech team. We'd overlooked adding this in the initial commit and it's something we rely on in our GitHub workflows.