Add Dependabot and a GitHub CI workflow (#1)
We like to keep our projects up to date to ensure we don't fall too far behind. If we did there are features we might miss out on, fixes we're not aware of, and security patches we should apply (though a test suite we do run them in our environments).

We also like to know that we haven't broken anything when we make a change. So, as a standard we have a [GitHub Workflow](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions) that runs just to confirm the project still builds and can run a specific CI test.

** Notes

Changes to allow the feature to be run in CI - We want to be able to run our CI feature in GitHub actions and only the CI feature. In doing this we've learnt that Cypress does an automatic check on the web service under test. If it cannot connect then it won't bother running the tests. This is a bit of a shame as our CI feature is intended to have no dependency on anything external but it's something we cannot workaround. Currently, the project is configured to default to selecting the 'local' environment if none is provided and this means it tries to connect to http://localhost:3000 in its startup check. So, rather than re-writing everything, the simplest solution was a CI .env file that uses a baseURL we know is simple and reliable (well done example.com!)

Add .nvmrc file - As a standard in our projects, we add an [.nvmrc](https://github.com/nvm-sh/nvm#nvmrc) to specify the version of Node being used. This helps ensure consistency across the tech team. We'd overlooked adding this in the initial commit and it's something we rely on in our GitHub workflows.

Fix `rm cypress/reports/json/*.json` if folder empty - Found when running in CI that if `rm` gets no match for what it's supposed to be deleting then it throws an error. Adding `-f` means force the removal but also means we won't get an error (non zero exit code) if the folder is empty.