Update a bill run scenario to demo multiple trans. (#4)
Before handing this project over to our QA team we want to ensure it contains a broad set of examples.

One of the key things we think we are missing is a scenario that adds more than one transaction. Most bill runs will have a few. Plus to test more complex scenarios we'll need to be able to different types of transactions to a bill run.

So, in this change, we are going to amend our generate bill run scenario to add a dynamic number of transactions. We hope the project is then in a good state to base our further testing on.