Minor housekeeping (#18)
This is just to tidy some things up we've noticed. Removal of extraneous spaces, some duplication, and moving the calculate charge endpoint to its own endpoints file.

** Notes

The endpoint being tested in `cypress/integration/calculate_charge_validation.feature` is actually `/calculate-charge`. To make the distinction between this and `/transactions` clearer we create a new endpoints class and move the existing methods to it.

By making `failOnStatusCode` parametised we can rely on just `calculate()` in `CalculateChargeEndpoints`.

And by telling the function `cypress/integration/calculate_charge_validation/calculate_validation_steps.js` whether a request is valid or not we can combine 'valid' and 'invalid into one method.