Update package-lock.json
Locally when we run `npm install` we are seeing the package-lock.json being modified. We think this is as a result of running the latest stable version of npm (8.1.3).

To avoid keep seeing it pop up as a change, and as we should be keeping our tools up to date this change commits the changes npm 8 is making.