Add full VSCode cucumber support
We use [cypress-cucumber-preprocessor](https://github.com/TheBrainFamily/cypress-cucumber-preprocessor) to allow us to write and organise our tests in a Behavior-Driven Development (BDD) style using [Cucumber](https://cucumber.io/).

On the preprocessor's README it talks about how you can setup VSCode to better support writing features. Typically we shy away from adding tool specific config to our code repos but we make the exception for VSCode as it's so prelific and everyone on the team uses it.

So, this change covers the basics of

- add the basic config to get [Cucumber Full Language Support](https://github.com/alexkrechik/VSCucumberAutoComplete) working with the features and step definitions in our project
- update the tests where necessary to remove any errors
- apply the auto-formatting the VSCode plugin uses to help us remain consistent