Add basic VSCode config
This allows vscode app specific config to be added to the project. The settings file covers the bare minimum needed to help the VSCode extension understand where to look for the step definitions. Any other config you can enable is more likely to be at the preference of each dev.