Add @wip tag support
Something we've been able to use in other test projects is the ability to tag 'work in progress' (WIP) scenarios and have the runner only run those. This is immensely helpful when building new tests or trying to figure out why one is failing.

The cypress-cucumber-preprocessor suggests @focus will do the trick but it doesn't seem to work for us. So, instead by adding additional run scripts to our `package.json` we can mimic the effect.