Fix failing calculate_charge_validation scenarios
The calculate-charge endpoint in [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) is now ready to be tested and we have some existing validation scenarios. However, a large number are failing. We believe this is just because of some additional changes made to the endpoint and not an issue with the endpoint itself.

This change covers correcting anything the scenarios are doing wrong so we can be certain what if anything is actually failing.