Start moving calculate charge to split features
To make it easier to separate out endpoints and features this change is the result of an experiment to see if we can move features into subfolders. Turns out we can!

So, it just contains some basic tests of the calculate charge endpoint to prove the idea works. Next we'll start migrating the validation feature and scenarios into the new folder structure.