Replace 'ruleset data item must be valid'
We've replaced the first validation test with a slightly different one. The existing one was actual just confirming that ruleset is a valid field and we can do that in the 'Data input items are mandatory' scenario.

Instead we focus on confirming that only `presroc` and `sroc` are accepted values and anything else is invalid. This then makes the scenario distinct from what the other is doing.