Update ruleset scenario to use a dataTable
Using a scenario outline with examples is great because it allows you to clearly run a scenario using a series of values.

The issue is when you have a large number of examples, plus multiple scenarios with large series of examples. Each example is a test in itself so this greatly expands the number of tests being run and makes the output very noisy.

What we really are trying to say with our scenario is that the API only accepts certain values for ruleset. We only really want to see that test listed once but know we've tried a number of values.

Using a mix of a [dataTable](https://github.com/TheBrainFamily/cypress-cucumber-preprocessor#data-table-parameters) and the [cypress `each`](https://docs.cypress.io/api/commands/each) command we can do this.