Update sroc charge fixture
This forms the basis for any SROC requests to the calculate charge endpoint. We were unable to get any requests to work with it in its current form. So, having compared to a working example we've updated the base fixture

- to default values to false with the intent to set to true in the tests
- use decimal values where the value can be a decimal
- update fields to match what the CM now expects

Note - We have exposed an error with the CM. If `compensationCharge` is false we get a 422 with a `compensationChargePercentage must be a string` error. It looks like the CM is always expecting this field to be populated in the rules service response. However, it only is if `compensationCharge` is true.