Replace existing default scenarios
Some fields if not included in the request will have a default value assigned to them. This migrates the existing tests to use the new datatable approach.

Note - We have also had to mark this scenario as `@ignore`. It is expecting to test certain fields in the response to confirm the default value was assigned. However, those fields are returned from the rules service and not the CM. So, at this time we're not sure how the scenario needs to be re-configured to confirm the default values are applied.