Migrate invalid combination scenarios
Note, we have a workaround to default `compensationCharge` to false as this is what we would normally get from the JSON fixture. However, until https://eaflood.atlassian.net/browse/CMEA-193 is resolved we need to do it within the test.

The AC's also discuss what combinations are valid so we also add a scenario for those. Again, we're blocked by the error with compensationCharge which is why it carries the @ignore tag.