Remove @ignore from sets defaults scenarios
Using the assumption that if the charge value, when all other things are equal, is the same when the property is not set and when set with the default value we can conclude that the CM defaults it as expected.

This means we can remove our TODO comment and the @ignore tag. We also include the one property we know presroc defaults.