Re-write combination tests
With the compensationCharge issue resolved we could then remove the @ignore tags from our combination scenarios. However, they still failed because they were not taking into account the relationship between twoPartTariff, compensationCharge and section127Agreement. Though the A/C's focus on pairs of combinations, there are times where the the remaining property needs to be set to something to allow the pair combination to be tested.

So, we've written them on the basis we always set all 3 properties and make assertions against what we expect when each combination of them is set.