Fix failing calculate_charge_validation scenarios (#23)
The calculate-charge endpoint in [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) is now ready to be tested and we have some existing validation scenarios. However, a large number are failing. We believe this is just because of some additional changes made to the endpoint and not an issue with the endpoint itself.

As part of this, we move the existing calculate charge feature to a folder and which then allows us to create separate features for different aspects of the calculate charge. This is a pattern we can bring to the other endpoints in time.

We also migrate the scenarios outlines to scenarios with data tables. Using data tables means we can still perform the same number of tests. However, Cypress groups them under a single scenario which makes the test output cleaner and easier to see.

We do this using a mix of a [dataTable](https://github.com/TheBrainFamily/cypress-cucumber-preprocessor#data-table-parameters) and the [cypress `each`](https://docs.cypress.io/api/commands/each) command.

We also had to update the SROC fixture. This forms the basis for any SROC requests to the calculate charge endpoint. We were unable to get any requests to work with it in its current form. So, having compared to a working example we updated the base fixture

- to default values to false with the intent to set to true in the tests
- use decimal values where the value can be a decimal
- update fields to match what the CM now expects