Add check for section126Factor decimal places
section126Factor only applies to presroc and has a limit on the number of decimal places that can be sent through. This is actually a limit of the rules service but unlike other values, it kind of blows up if we send it something with more than 3 decimal places.

So, we have the check in the CM and this scenario ensures the check is working.