All tests passing (with a but)
Just wanted to record at this point in the process all tests are passing apart from 1 in calculate charge and one in transactions. Both depend on using a supported source name that currently has an invalid hyphen. This was fixed in the CM in https://github.com/DEFRA/sroc-charging-module-api/pull/652 but as yet has not been corrected in the tests.