Update delete bill run endpoint to v3
https://eaflood.atlassian.net/browse/CMEA-233

This was intended to be about migrating the existing postman test to this project but then we found it was already done!

So, the only change we need to make is to update the endpoints file to use `/v3` instead of `/v2` as we have done with `/approve` and `/send`.