Add scenarios
We've not covered every single possible status a bill run can be in, mainly because the transition ones can be fleeting and would be unreliable to test against. We have unit tests for those.

But we have covered the main ones and hopefully made it clear how we expect the rebill function to behave depending on whether the destination bill run is editable or not.

Note - 'Destination bill run cannot be approved' currently fails because we as yet have not merged the fix.