Remove redundant bill run endpoint method (#40)
Spotted whilst working on something else we have

```javascript
  static view (id, failOnStatusCode = true) {
    return cy
      .api({
        method: 'GET',
        url: `/v3/wrls/bill-runs/${id}`,
        failOnStatusCode,
        headers: {
          'content-type': 'application/json',
          accept: 'application/json',
          Authorization: `Bearer ${Cypress.env('token')}`
        }
      })
  }

  static viewDeleted (id) {
    return cy
      .api({
        method: 'GET',
        url: `/v2/wrls/bill-runs/${id}`,
        failOnStatusCode: false,
        headers: {
          'content-type': 'application/json',
          accept: 'application/json',
          Authorization: `Bearer ${Cypress.env('token')}`
        }
      })
  }
```

They are essentially the same. The only difference is that `viewDeleted()` has been hardcoded to not fail when you attempt to view a bill run that does not exist. You can get the same result by setting the `failOnStatusCode` param of `view()` to false which makes `viewDeleted()` redundant.