Housekeeping - file name conventions (#38)
The first iteration of this project was our first attempt at using Cypress for API testing. We are also relatively new to the framework.

As such we have learnt more about how exactly we can structure our features and steps and that has resulted in some inconsistency in how we name things.

This change takes the opportunity to bring things in line with how we currently intend to name our files and folders.