Update supported source names with bad hyphens
https://github.com/DEFRA/sroc-charging-module-api/pull/652

It was spotted in the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) that our reference list of supported source names had a mix of hyphens and dashs. The list had been copied from an Excel spreadsheet which we believe had auto-changed some of the hyphens to en dashes without us noticing.

We've used the same list in the tests here but now we have corrected them in the API the scenarios are failing. This updates any supported source names with dashes to use hyphens instead.