Add scenario for new rebill invoices (#41)
https://github.com/DEFRA/sroc-charging-module-api/pull/658

We spotted whilst working on the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) that we were not checking the status of the bill run an invoice would be re-billed onto. We have checks that for the status of the bill run the original is linked to. But we were supposed to be checking that the bill run the invoice would be rebilled onto should be in an 'editable' state. Currently, this means its status should be either `initialised` or `generated`.

If the bill run is in any other state we should be rejecting the request. We'll soon have that [fixed](https://github.com/DEFRA/sroc-charging-module-api/pull/658). But as we overlooked this in both unit and acceptance testing we're making sure we have it covered this time.

** Notes

We've not covered every single possible status a bill run can be in, mainly because the transition ones can be fleeting and would be unreliable to test against. We have unit tests for those.

But we have covered the main ones and hopefully made it clear how we expect the rebill function to behave depending on whether the destination bill run is editable or not.