Fix environment variable loading
We currently specify the environment the Cypress tests are running in and use `cypress-dotenv` to load the appropriate file from `environments`. We don't want to do this if the tests are running in the `docker` environment as the env vars will already be present. We therefore amend our Cypress plugins file to not load a file if the environment is `docker`