Use multi-stage builds
Use multi-stage builds to split up the steps in the build process. Also, hoped to reduce the final image size but still seeing 1.2GB.

Other thing to note is we tried setting a home directory but this broke `npm install`

- https://github.com/cypress-io/cypress/issues/3081
- https://stackoverflow.com/questions/55151786/getting-error-eacces-permission-denied-when-i-install-cypress

There are workarounds for it but we'll come to those later once we have squeezed all we can out of the size.