Verify Cypress as part of build
On a fresh install of Cypress when you first use it Cypress kicks off a verify process. Because of the way containers work this would mean you'd have to wait for the process to complete each time you start a new container.

By adding it to the Dockerfile we can verify cypress as part of the build and therefore skip this step whenever you come to start a new container.