Some updates to tasks.json
First key one is the image name. We've changed it to match the project as this is a convention we have stuck with in other projects.

We also use the tag `latest` as that is often used as the default for scratch built images and is the default tag applied if you don't provide one.

We rename the build with no cache task as `REBUILD`. Though Docker Compose, based in other projects we have named the task that builds docker images from scratch `REBUILD`. This matches in with that convention.

Finally, we've added a `DOWN` task. Either once the tests finish or if you kill the container process the container still remains. This means the next time you call `docker run` the previous container will be restarted.

Things may have changed in the interim so to ensure the tests behave as expected it is prudent to remove any stopped containers for the project.