Fix apostrophes for windows
One of the team who uses window ran into a problem when using the **wip** functionaity of the project. Even though they were using the tag the relevant scenario was not being run.

After some investigation we tracked it down to how windows was handling the call we have in `package.json`. Using double quotes rather than single seemed to be a requirement to get it to work on the windows machine.

So, this change switchs the single quotes for double quotes in the package.json. We have tested on our non-windows machines and the calls still work there.