Avoid copying old reports and OSX files into image
Spotted that we could be copying old reports generated by a user into the image built plus local OSX specific files. It's not a major issue; we expect the official Docker image to be built with GitHub actions in an environment where these files wouldn't exist.

But just to be clean and you never know, we might need to build the image locally if the CI breaks, we update `.dockerignore` to ensure these files don't get copied in.