Set a working directory in the image
Prior to this change we were just installing everything into the root of the image. This would be the same as installing a new app into `c:\` on Windows or `/` on Linux.

We don't want our files mixed up with the root OS directories. So, we set a WORKDIR. But this then threw up a Cypress issue when it came to installing it

```
Cypress cannot write to the cache directory due to file permissions
Failed to access /root/.cache/Cypress:
```

This issue is covered here https://github.com/cypress-io/cypress/issues/3081. Some workarounds are mentioned plus a review of what Cypress actually do in their Dockerfiles highlighted

```
*# https://github.com/cypress-io/cypress-docker-images/blob/master/base/17.3.0/Dockerfile
*# allow installing when the main user is root
ENV npm_config_unsafe_perm true
```

```
*# https://github.com/cypress-io/cypress-docker-images/blob/master/included/3.2.0/Dockerfile
RUN npm config -g set user $(whoami)
```

Our testing showed we only need to set the user in npm's config and then it would install without error and still run the tests. So, until we see otherwise this is what we'l go with.