Put pack all dependencies
In hindsight, their docs say these are the dependencies and it matches with what we see in some of their own Dockerfiles. Who are we to start removing some even if it appears to still work. It certainly makes no difference to the finished size.

So, this restore the original list of OS dependencies Cypress says it needs.