Run tests in Docker (#48)
* Set up `.env` for Docker

We will supply the required environment variables for Docker. For running locally, this means creating a `.env` file and passing it in the `docker run` command. We therefore create `.env.example` and add `.env` to `.gitignore` to prevent us checking in credentials.

* Create `Dockerfile`

We create our `Dockerfile`. This sets up the Docker image then runs the `cy:run:docker` script, which we add to `package.json`

* Fix environment variable loading

We currently specify the environment the Cypress tests are running in and use `cypress-dotenv` to load the appropriate file from `environments`. We don't want to do this if the tests are running in the `docker` environment as the env vars will already be present. We therefore amend our Cypress plugins file to not load a file if the environment is `docker`

* Create VS Code tasks

We create a `.vscode/tasks.json` file with tasks to build and run the Docker image

* Use multi-stage builds

Use multi-stage builds to split up the steps in the build process. Also, hoped to reduce the final image size but still seeing 1.2GB.

Other thing to note is we tried setting a home directory but this broke `npm install`

- https://github.com/cypress-io/cypress/issues/3081
- https://stackoverflow.com/questions/55151786/getting-error-eacces-permission-denied-when-i-install-cypress

There are workarounds for it but we'll come to those later once we have squeezed all we can out of the size.

* Removed 2 dependencies from the base

Using https://hub.docker.com/r/cypress/base/dockerfile as a base started removing items that were not listed. Managed to remove 2 but anything else caused an error.

I can see where the inclusion of the others came from when you look at something like https://github.com/cypress-io/cypress-docker-images/blob/master/base/17.3.0/Dockerfile and a whole bunch more.

Again, hoped to see a dint in the size but `docker images` is stubbornly showing 1.2GB still!

* Change environments/ folder handling

After a test confirmed the docker image doesn't need the `environments/` folder in order to run the tests. So, we no longer bother to create one in the Dockerfile.

* Verify Cypress as part of build

On a fresh install of Cypress when you first use it Cypress kicks off a verify process. Because of the way containers work this would mean you'd have to wait for the process to complete each time you start a new container.

By adding it to the Dockerfile we can verify cypress as part of the build and therefore skip this step whenever you come to start a new container.

* Some updates to tasks.json

First key one is the image name. We've changed it to match the project as this is a convention we have stuck with in other projects.

We also use the tag `latest` as that is often used as the default for scratch built images and is the default tag applied if you don't provide one.

We rename the build with no cache task as `REBUILD`. Though Docker Compose, based in other projects we have named the task that builds docker images from scratch `REBUILD`. This matches in with that convention.

Finally, we've added a `DOWN` task. Either once the tests finish or if you kill the container process the container still remains. This means the next time you call `docker run` the previous container will be restarted.

Things may have changed in the interim so to ensure the tests behave as expected it is prudent to remove any stopped containers for the project.

* Avoid copying old reports and OSX files into image

Spotted that we could be copying old reports generated by a user into the image built plus local OSX specific files. It's not a major issue; we expect the official Docker image to be built with GitHub actions in an environment where these files wouldn't exist.

But just to be clean and you never know, we might need to build the image locally if the CI breaks, we update `.dockerignore` to ensure these files don't get copied in.

* Set a working directory in the image

Prior to this change we were just installing everything into the root of the image. This would be the same as installing a new app into `c:\` on Windows or `/` on Linux.

We don't want our files mixed up with the root OS directories. So, we set a WORKDIR. But this then threw up a Cypress issue when it came to installing it

```
Cypress cannot write to the cache directory due to file permissions
Failed to access /root/.cache/Cypress:
```

This issue is covered here https://github.com/cypress-io/cypress/issues/3081. Some workarounds are mentioned plus a review of what Cypress actually do in their Dockerfiles highlighted

```
*# https://github.com/cypress-io/cypress-docker-images/blob/master/base/17.3.0/Dockerfile
*# allow installing when the main user is root
ENV npm_config_unsafe_perm true
```

```
*# https://github.com/cypress-io/cypress-docker-images/blob/master/included/3.2.0/Dockerfile
RUN npm config -g set user $(whoami)
```

Our testing showed we only need to set the user in npm's config and then it would install without error and still run the tests. So, until we see otherwise this is what we'l go with.

* Put pack all dependencies

In hindsight, their docs say these are the dependencies and it matches with what we see in some of their own Dockerfiles. Who are we to start removing some even if it appears to still work. It certainly makes no difference to the finished size.

So, this restore the original list of OS dependencies Cypress says it needs.

* Add comments to the Dockerfile

* Update `.env.example`

We left `http://` out of `CYPRESS_BASE_URL` -- we put it in to make it clear that it's needed

* Add Docker instructions to `README.md`

* Fix small `README.md` typo

We noticed that we were referring to the environment example `.env.example` -- the usual name for an example `.env` file -- but we have actually named it `.example.env`. Since this name is actually a better example of how we name it (with the environment first then `.env`) we amend `README.md` to match

* Small `README.md` tweaks

- We noticed that we had not been consistent about linking to files and folders in the repo. We therefore add links in all instances.
- We add in missing backticks (eg. where a file was being linked).
- We make some swall tweaks to the wording of the Docker instructions.

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>